Fixed a compile error that did not show up in Eclipse :(
git-svn-id: https://cdk.svn.sourceforge.net/svnroot/cdk/trunk/cdk@5856 eb4e18e3-b210-0410-a6ab-dec725e4b171

+++ src/org/openscience/cdk/test/qsar/descriptors/molecular/BCUTDescriptorTest.java
@@ -69,7 +69,7 @@ public class BCUTDescriptorTest extends CDKTestCase {
         assertNotNull(retval);
         /* System.out.println("Num ret = "+retval.size()); */
         for (int i = 0; i < retval.size(); i++) {
-            assertTrue(Math.abs(0.0 - ((Double)retval.get(i)).doubleValue()) > 0.0000001);
+            assertTrue(Math.abs(0.0 - retval.get(i)) > 0.0000001);
         }
 
         /*
